+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## unreleased
 
+## 0.3.3, 2026-08-30
+
  - feat(upgrade): added `upgrade` alias for `update`
 - feat(upgrade): default `--max-concurrency` raised from `2` to `5`
 - feat(reboot): `-t`/`--terminal` opens a terminal in each rebooted VM